Check for reads size test
