+++ tests/test_calculate_delta.py
@@ -1,0 +1,82 @@
+from fai.contracts import Read
+from fai.logic import calculate_delta
+
+
+def test_calculate_delta():
+    gene_dict = {
+        "gene1": "TAACAAACAAAACAAATACAATGGTCAAATTAACTTCAATCGCCGCTGGTGTCGCTGCCATCGCTGCTACTGCTTCTGCAACCACCACTCTAGCTCAATCTGACGAAAGAGTCAACTTGGTGGAATTGGGTGTCTACGTCTCTGATATCAGAGCTCACTTAGCCCAATACTACATGTTCCAAGCCGCCCACCCAACTGAAACCTACCCAGTCGAAGTTGCTGAAGCCGTTTTCAACTACGGTGACTTCACCACCATGTTGACCGGTATTGCTCCAGACCAAGTGACCAGAATGATCACCGGTGTTCCATGGTACTCCAGCAGATTAAAGCCAGCCATCTCCAGTGCTCTATCCAAGGACGGTATCTACACTATCGCAAACTAGAGACAAACGCCAATTTTACG",  # noqa: E501
+        "gene2": "CATGAGGTCCCACTGTATGTATGTGCCCTCGCACAGTTTTGCTCATCATAAACATTAATCATTGGTTTTATGATAAGAATATCGTAAGAATTATTTTAACTTTCAGATTAGATTCCGGCCATATTTCTGACATTTGTTTTATTAATAAAAATTTGGCGAATGCTTTGATAACTGCTGATATTTCATTGTTGAAAAGGCATGATATAAGATGCACAAAGTATATTATAACTTATTATCAGAGATATAGGAACAAAGAAAAAGGGAAATTCATATCTCTATGTAAAAATACCATTATTTCCTCTTCTGTTTAATATGCTGTCCTTAATTTTCA",  # noqa: E501
+        "gene3": "GTACAATAAAAGCCCCAAAAATGTCTCTGGCACATTATTGTTTACTACTAGCCATCGTCACATTGCTGGGATTAACTAATGTTGTCTCTGCGACTACAGCGGCATGCCTGCCAGCAAACTCAAGGAAGAATGGTATGAATGTAAACTTTTACCAGTATTCATTGAGAGATTCCTCCACATATTCGAATGCAGCATATATGGCTTATGGATATGCCTCAAAAACTAAACTGGGTTCTGTCGGAGGACAAACTGATATCTCGATTGATTATAATATTCCTTGTGTTAGTTCATCAGGCACATTTCCTTGTCCTCAAGAAGATTTATATGGTAATTGGGGATGCAAAGGAATTGGTGCTTGTTCTAATAATCCAATAATTGCATACTGGAGTACTGATTTATTTGGTTTCTATACTACCCCAACAAACGTAACCCTAGAAATGACAGGTTATTTTTTACCACCACAGACGGGTTCTTACACATTCAAGTTTGCTACAGTTGACGACTCTGCAATTCTATCAGTCGGTGGTAGCATTGCGTTCGAATGTTGTGCACAAGAACAACCTCCCATCACGTCGACTAACTTCACCATCAATGGTATCAAGCCATGGAATGGAAGTCCCCCTGATAATATTACAGGGACTGTCTACATGTATGCTGGTTTCTATTATCCAATGAAGATTGTTTACTCAAATGCCGTTGCCTGGGGTACACTTCCAATTAGTGTGACACTACCAGATGGCACTACCGTTAGTGATGACTTTGAAGGGTACGTATATACTTTTGACAACAATCTAAGCCAGCCAAACTGTACCATTCCAGACCCTTCAAATTATACTGTCAGTACTACCATAACTACAACGGAACCATGGACCGGTACTTTCACTTCTACATCTACTGAAATGACCACCGTCACCGGTACCAACGGCGTTCCAACTGACGAAACCGTCATTGTCATCAGAACTCCAACAACTGCTAGCACCATCATAACTACAACTGAGCCATGGAACAGCACTTTTACCTCTACTTCTACCGAATTGACCACAGTCACTGGCACCAATGGTGTACGAACTGACGAAACCATCATTGTAATCAGAACACCAACAACAGCCACTACTGCCATAACTACAACTGAGCCATGGAACAGCACTTTTACCTCTACTTCTACCGAATTGACCACAGTCACCGGTACCAATGGTTTGCCAACTGATGAGACCATCATTGTCATCAGAACACCAACAACAGCCACTACTGCCATGACTACAACTCAGCCATGGAACGACACTTTTACCTCTACTTCTACCGAATTGACCACAGTCACCGGTACCAATGGTTTGCCAACTGATGAGACCATCATTGTCATCAGAACACCAACAACAGCCACTACTGCCATGACTACAACTCAGCCATGGAACGACACTTTTACCTCTACTTCTACCGAATTGACCACAGTCACCGGTACCAATGGTTTGCCAACTGATGAGACCATCATTGTCATCAGAACACCAACAACAGCCACTACTGCCATGACTACAACTCAGCCATGGAACGACACTTTTACCTCTACATCCACTGAAATCACCACCGTCACCGGTACCAATGGTTTGCCAACTGATGAGACCATCATTGTCATCAGAACACCAACAACAGCCACTACTGCCATGACTACAACTCAGCCATGGAACGACACTTTTACCTCTACATCCACTGAAATGACCACCGTCACCGGTACCAACGGTTTGCCAACTGATGAAACCATCATTGTCATCAGAACACCAACAACAGCCACTACTGCCATAACTACAACTGAGCCATGGAACAGCACTTTTACCTCTACATCCACTGAAATGACCACCGTCACCGGTACCAACGGTTTGCCAACTGATGAAACCATCATTGTCATCAGAACACCAACAACAGCCACTACTGCCATAACTACAACTCAGCCATGGAACGACACTTTTACCTCTACATCCACTGAAATGACCACCGTCACCGGTACCAACGGTTTGCCAACTGATGAAACCATCATTGTCATCAGAACACCAACAACAGCCACTACTGCCATGACTACAACTCAGCCATGGAACGACACTTTTACCTCTACATCCACTGAAATCACCACCGTCACCGGTACCAACGGTTTGCCAACTGATGAGACCATCATTGTCATCAGAACACCAACAACAGCCACTACTGCCATGACTACAACTCAGCCATGGAACGACACTTTTACCTCTACATCCACTGAAATGACCACCGTCACCGGTACCAACGGCGTTCCAACTGACGAAACCGTCATTGTCATCAGAACTCCAACTAGTGAAGGTCTAATCAGCACCACCACTGAACCATGGACTGGTACTTTCACCTCTACATCCACTGAGATGACCACCGTCACCGGTACTAACGGTCAACCAACTGACGAAACCGTGATTGTTATCAGAACTCCAACCAGTGAAGGTTTGGTTACAACTACAACCGAGCCATGGACCGGTACTTTCACCTCTACATCTACTGAGATGACCACCATCACTGGAACCAACGGTCAACCAACTGATGAAACTGTCATTATTGTCAAAACTCCAACTACTGCCATCTCATCCAGTTTGTCATCTTCTTCAGGACAAATCACCAGCTTTATCACGTCTGCGCGTCCAATTATTACCCCATTCTATCCTAGCAATGGAACTTCTGTGATTTCCTCCTCAGTAATTTCTTCCTCAGACACTTCTTCTCTAGTCATTTCTTCCTCAGTCACTTCTTCTCTAGTCACTTCTTCTCCAGTCATTTCTTCTTCATTCATTTCTTCCCCTGTCATTTCTTCTACAACAACCTCCGCTTCTATACTCTCTGAATCATCTAAATCATCCGTCATTCCAACCAGTAGTTCCACCTCTGGTTCTTCTGAGAGCGAAACGGGTTCAGCTAGTTCTGCCTCTTCTTCCTCTTCTATCTCTTCTGAATCACCAAAGTCTACATATTCGTCTTCATCATTACCACCTGTTACCAGTGCAACAACAAGTCAGGAAATTACTTCTTCATTACCACCTGTTACCACTACAAAAACGAGCGAACAAACCACTTTGGTTACCGTGACATCCTGCGAATCTCATGTGTGCACTGAATCTATCTCCTCTGCGATTGTTTCCACGGCCACCGTTACTGTTAGCGGTGCCACAACAGAGTATACCACATGGTGCCCTATTTCTACCACAGAGATAACAAAGCAAACTACGGAGACAACAAAGCAAACCAAGGGGACAACAGAGCAAACCACAGAAACAACAAAACAAACCACAGTAGTTACAATTTCTTCTTGTGAATCTGACGTATGCTCTAAGACTGCTTCTCCAGCCATTGTATCTACAAGCACTGCTACTATTAATGGCGTTACCACAGAATACACAACATGGTGTCCTATTTCCACCACAGAATCGAAGCAACAAACTACGCTAGTTACTGTTACTTCCTGCGGATCTGGTGTGTGTTCCGAAACTACTTCACCTGCCATTGTTTCGACGGCCACGGCTACTGTGAATGATGTTGTTACGGTCTATTCTACATGGAGGCCACAGACTACGAATGAACAGTCTGTCAGCTCTAAAATGAACAGTGCTACCAGTGAGACAACAACCAATACTGGAGCTGCTGAGACAACTACCAGTACTGGAGCTGCTGAGACGAAAACAGTAGTCACCTCTTCAATTTCAAGATTCAATCATGCTGAAACACAGACGGCTTCCGCGACCGATGTGATTGGTCACAGCAGTAGTGTTGTTTCTGTATCCGAAACTGGCAACACCAAGAGTCTAACAAGTTCCGGGTTGAGTACTATGTCGCAACAGCCTCGTAGCACACCAGCAAGTAGCATGGTAGGATCTAGTACAGCTTCTTTAGAAATTTCAACGTATGCTGGCAGTGCCAACAGCTTACTGGCCGGTAGTGGTTTAAGTGTCTTCATTGCGTCCTTATTGCTGGCAATTATTTAATAAAATTCGCGTTCTTTTTA",  # noqa: E501
+    }
+
+    start_codon_index = 20
+    read_size = 30
+    reads_spacing = 3
+
+    expected_delta = 15
+
+    reads = [
+        Read(
+            gene, start_codon_index - expected_delta + offset * reads_spacing, read_size
+        )
+        for gene in gene_dict
+        for offset in range(
+            0,
+            (len(gene_dict[gene]) - read_size - (start_codon_index - expected_delta))
+            // reads_spacing
+            + 1,
+        )
+    ]
+
+    delta = calculate_delta(reads, gene_dict)
+
+    assert delta == expected_delta
+
+    expected_delta = 14
+
+    reads = [
+        Read(
+            gene, start_codon_index - expected_delta + offset * reads_spacing, read_size
+        )
+        for gene in gene_dict
+        for offset in range(
+            0,
+            (len(gene_dict[gene]) - read_size - (start_codon_index - expected_delta))
+            // reads_spacing
+            + 1,
+        )
+    ]
+
+    assert delta == expected_delta
+
+    expected_delta = 16
+
+    reads = [
+        Read(
+            gene, start_codon_index - expected_delta + offset * reads_spacing, read_size
+        )
+        for gene in gene_dict
+        for offset in range(
+            0,
+            (len(gene_dict[gene]) - read_size - (start_codon_index - expected_delta))
+            // reads_spacing
+            + 1,
+        )
+    ]
+
+    assert delta == expected_delta
+
+
+def test_calculate_delta_with_unmatching_size_reads():
+    gene_dict = {
+        "gene1": "TAACAAACAAAACAAATACAATGGTCAAATTAACTTCAATCGCCGCTGGTGTCGCTGCCATCGCTGCTACTGCTTCTGCAACCACCACTCTAGCTCAATCTGACGAAAGAGTCAACTTGGTGGAATTGGGTGTCTACGTCTCTGATATCAGAGCTCACTTAGCCCAATACTACATGTTCCAAGCCGCCCACCCAACTGAAACCTACCCAGTCGAAGTTGCTGAAGCCGTTTTCAACTACGGTGACTTCACCACCATGTTGACCGGTATTGCTCCAGACCAAGTGACCAGAATGATCACCGGTGTTCCATGGTACTCCAGCAGATTAAAGCCAGCCATCTCCAGTGCTCTATCCAAGGACGGTATCTACACTATCGCAAACTAGAGACAAACGCCAATTTTACG",  # noqa: E501
+    }
+
+    reads = [Read("gene1", 20, 25), Read("gene1", 50, 30), Read("gene1", 80, 35)]
+
+    try:
+        calculate_delta(reads, gene_dict)
+    except ValueError as e:
+        assert str(e) == "All reads must have the same size."
+    else:
+        assert False, "Expected ValueError was not raised."
