+++ tests/test_logic.py
@@ -1,5 +1,5 @@
 from fai.contracts import Read
-from fai.logic import add_flankers_to_reads, calculate_delta, get_flankers
+from fai.logic import add_flankers_to_reads, get_flankers
 
 
 def test_add_flankers_to_read():
@@ -67,68 +67,3 @@ def test_get_flankers_gene_overflow():
         )
     else:
         assert False, "Expected ValueError was not raised."
-
-
-def test_calculate_delta():
-    fasta_dict = {
-        "gene1": "TAACAAACAAAACAAATACAATGGTCAAATTAACTTCAATCGCCGCTGGTGTCGCTGCCATCGCTGCTACTGCTTCTGCAACCACCACTCTAGCTCAATCTGACGAAAGAGTCAACTTGGTGGAATTGGGTGTCTACGTCTCTGATATCAGAGCTCACTTAGCCCAATACTACATGTTCCAAGCCGCCCACCCAACTGAAACCTACCCAGTCGAAGTTGCTGAAGCCGTTTTCAACTACGGTGACTTCACCACCATGTTGACCGGTATTGCTCCAGACCAAGTGACCAGAATGATCACCGGTGTTCCATGGTACTCCAGCAGATTAAAGCCAGCCATCTCCAGTGCTCTATCCAAGGACGGTATCTACACTATCGCAAACTAGAGACAAACGCCAATTTTACG",  # noqa: E501
-        "gene2": "CATGAGGTCCCACTGTATGTATGTGCCCTCGCACAGTTTTGCTCATCATAAACATTAATCATTGGTTTTATGATAAGAATATCGTAAGAATTATTTTAACTTTCAGATTAGATTCCGGCCATATTTCTGACATTTGTTTTATTAATAAAAATTTGGCGAATGCTTTGATAACTGCTGATATTTCATTGTTGAAAAGGCATGATATAAGATGCACAAAGTATATTATAACTTATTATCAGAGATATAGGAACAAAGAAAAAGGGAAATTCATATCTCTATGTAAAAATACCATTATTTCCTCTTCTGTTTAATATGCTGTCCTTAATTTTCA",  # noqa: E501
-        "gene3": "GTACAATAAAAGCCCCAAAAATGTCTCTGGCACATTATTGTTTACTACTAGCCATCGTCACATTGCTGGGATTAACTAATGTTGTCTCTGCGACTACAGCGGCATGCCTGCCAGCAAACTCAAGGAAGAATGGTATGAATGTAAACTTTTACCAGTATTCATTGAGAGATTCCTCCACATATTCGAATGCAGCATATATGGCTTATGGATATGCCTCAAAAACTAAACTGGGTTCTGTCGGAGGACAAACTGATATCTCGATTGATTATAATATTCCTTGTGTTAGTTCATCAGGCACATTTCCTTGTCCTCAAGAAGATTTATATGGTAATTGGGGATGCAAAGGAATTGGTGCTTGTTCTAATAATCCAATAATTGCATACTGGAGTACTGATTTATTTGGTTTCTATACTACCCCAACAAACGTAACCCTAGAAATGACAGGTTATTTTTTACCACCACAGACGGGTTCTTACACATTCAAGTTTGCTACAGTTGACGACTCTGCAATTCTATCAGTCGGTGGTAGCATTGCGTTCGAATGTTGTGCACAAGAACAACCTCCCATCACGTCGACTAACTTCACCATCAATGGTATCAAGCCATGGAATGGAAGTCCCCCTGATAATATTACAGGGACTGTCTACATGTATGCTGGTTTCTATTATCCAATGAAGATTGTTTACTCAAATGCCGTTGCCTGGGGTACACTTCCAATTAGTGTGACACTACCAGATGGCACTACCGTTAGTGATGACTTTGAAGGGTACGTATATACTTTTGACAACAATCTAAGCCAGCCAAACTGTACCATTCCAGACCCTTCAAATTATACTGTCAGTACTACCATAACTACAACGGAACCATGGACCGGTACTTTCACTTCTACATCTACTGAAATGACCACCGTCACCGGTACCAACGGCGTTCCAACTGACGAAACCGTCATTGTCATCAGAACTCCAACAACTGCTAGCACCATCATAACTACAACTGAGCCATGGAACAGCACTTTTACCTCTACTTCTACCGAATTGACCACAGTCACTGGCACCAATGGTGTACGAACTGACGAAACCATCATTGTAATCAGAACACCAACAACAGCCACTACTGCCATAACTACAACTGAGCCATGGAACAGCACTTTTACCTCTACTTCTACCGAATTGACCACAGTCACCGGTACCAATGGTTTGCCAACTGATGAGACCATCATTGTCATCAGAACACCAACAACAGCCACTACTGCCATGACTACAACTCAGCCATGGAACGACACTTTTACCTCTACTTCTACCGAATTGACCACAGTCACCGGTACCAATGGTTTGCCAACTGATGAGACCATCATTGTCATCAGAACACCAACAACAGCCACTACTGCCATGACTACAACTCAGCCATGGAACGACACTTTTACCTCTACTTCTACCGAATTGACCACAGTCACCGGTACCAATGGTTTGCCAACTGATGAGACCATCATTGTCATCAGAACACCAACAACAGCCACTACTGCCATGACTACAACTCAGCCATGGAACGACACTTTTACCTCTACATCCACTGAAATCACCACCGTCACCGGTACCAATGGTTTGCCAACTGATGAGACCATCATTGTCATCAGAACACCAACAACAGCCACTACTGCCATGACTACAACTCAGCCATGGAACGACACTTTTACCTCTACATCCACTGAAATGACCACCGTCACCGGTACCAACGGTTTGCCAACTGATGAAACCATCATTGTCATCAGAACACCAACAACAGCCACTACTGCCATAACTACAACTGAGCCATGGAACAGCACTTTTACCTCTACATCCACTGAAATGACCACCGTCACCGGTACCAACGGTTTGCCAACTGATGAAACCATCATTGTCATCAGAACACCAACAACAGCCACTACTGCCATAACTACAACTCAGCCATGGAACGACACTTTTACCTCTACATCCACTGAAATGACCACCGTCACCGGTACCAACGGTTTGCCAACTGATGAAACCATCATTGTCATCAGAACACCAACAACAGCCACTACTGCCATGACTACAACTCAGCCATGGAACGACACTTTTACCTCTACATCCACTGAAATCACCACCGTCACCGGTACCAACGGTTTGCCAACTGATGAGACCATCATTGTCATCAGAACACCAACAACAGCCACTACTGCCATGACTACAACTCAGCCATGGAACGACACTTTTACCTCTACATCCACTGAAATGACCACCGTCACCGGTACCAACGGCGTTCCAACTGACGAAACCGTCATTGTCATCAGAACTCCAACTAGTGAAGGTCTAATCAGCACCACCACTGAACCATGGACTGGTACTTTCACCTCTACATCCACTGAGATGACCACCGTCACCGGTACTAACGGTCAACCAACTGACGAAACCGTGATTGTTATCAGAACTCCAACCAGTGAAGGTTTGGTTACAACTACAACCGAGCCATGGACCGGTACTTTCACCTCTACATCTACTGAGATGACCACCATCACTGGAACCAACGGTCAACCAACTGATGAAACTGTCATTATTGTCAAAACTCCAACTACTGCCATCTCATCCAGTTTGTCATCTTCTTCAGGACAAATCACCAGCTTTATCACGTCTGCGCGTCCAATTATTACCCCATTCTATCCTAGCAATGGAACTTCTGTGATTTCCTCCTCAGTAATTTCTTCCTCAGACACTTCTTCTCTAGTCATTTCTTCCTCAGTCACTTCTTCTCTAGTCACTTCTTCTCCAGTCATTTCTTCTTCATTCATTTCTTCCCCTGTCATTTCTTCTACAACAACCTCCGCTTCTATACTCTCTGAATCATCTAAATCATCCGTCATTCCAACCAGTAGTTCCACCTCTGGTTCTTCTGAGAGCGAAACGGGTTCAGCTAGTTCTGCCTCTTCTTCCTCTTCTATCTCTTCTGAATCACCAAAGTCTACATATTCGTCTTCATCATTACCACCTGTTACCAGTGCAACAACAAGTCAGGAAATTACTTCTTCATTACCACCTGTTACCACTACAAAAACGAGCGAACAAACCACTTTGGTTACCGTGACATCCTGCGAATCTCATGTGTGCACTGAATCTATCTCCTCTGCGATTGTTTCCACGGCCACCGTTACTGTTAGCGGTGCCACAACAGAGTATACCACATGGTGCCCTATTTCTACCACAGAGATAACAAAGCAAACTACGGAGACAACAAAGCAAACCAAGGGGACAACAGAGCAAACCACAGAAACAACAAAACAAACCACAGTAGTTACAATTTCTTCTTGTGAATCTGACGTATGCTCTAAGACTGCTTCTCCAGCCATTGTATCTACAAGCACTGCTACTATTAATGGCGTTACCACAGAATACACAACATGGTGTCCTATTTCCACCACAGAATCGAAGCAACAAACTACGCTAGTTACTGTTACTTCCTGCGGATCTGGTGTGTGTTCCGAAACTACTTCACCTGCCATTGTTTCGACGGCCACGGCTACTGTGAATGATGTTGTTACGGTCTATTCTACATGGAGGCCACAGACTACGAATGAACAGTCTGTCAGCTCTAAAATGAACAGTGCTACCAGTGAGACAACAACCAATACTGGAGCTGCTGAGACAACTACCAGTACTGGAGCTGCTGAGACGAAAACAGTAGTCACCTCTTCAATTTCAAGATTCAATCATGCTGAAACACAGACGGCTTCCGCGACCGATGTGATTGGTCACAGCAGTAGTGTTGTTTCTGTATCCGAAACTGGCAACACCAAGAGTCTAACAAGTTCCGGGTTGAGTACTATGTCGCAACAGCCTCGTAGCACACCAGCAAGTAGCATGGTAGGATCTAGTACAGCTTCTTTAGAAATTTCAACGTATGCTGGCAGTGCCAACAGCTTACTGGCCGGTAGTGGTTTAAGTGTCTTCATTGCGTCCTTATTGCTGGCAATTATTTAATAAAATTCGCGTTCTTTTTA",  # noqa: E501
-    }
-
-    start_codon_index = 20
-    read_size = 30
-    reads_spacing = 3
-
-    expected_delta = 15
-
-    reads = [
-        Read(
-            gene, start_codon_index - expected_delta + offset * reads_spacing, read_size
-        )
-        for gene in fasta_dict
-        for offset in range(
-            0,
-            (len(fasta_dict[gene]) - read_size - (start_codon_index - expected_delta))
-            // reads_spacing
-            + 1,
-        )
-    ]
-
-    delta = calculate_delta(reads)
-
-    assert delta == expected_delta
-
-    expected_delta = 14
-
-    reads = [
-        Read(
-            gene, start_codon_index - expected_delta + offset * reads_spacing, read_size
-        )
-        for gene in fasta_dict
-        for offset in range(
-            0,
-            (len(fasta_dict[gene]) - read_size - (start_codon_index - expected_delta))
-            // reads_spacing
-            + 1,
-        )
-    ]
-
-    assert delta == expected_delta
-
-    expected_delta = 16
-
-    reads = [
-        Read(
-            gene, start_codon_index - expected_delta + offset * reads_spacing, read_size
-        )
-        for gene in fasta_dict
-        for offset in range(
-            0,
-            (len(fasta_dict[gene]) - read_size - (start_codon_index - expected_delta))
-            // reads_spacing
-            + 1,
-        )
-    ]
-
-    assert delta == expected_delta
